Fix: Wrong roll number value on repetitive fast clicks
When a user is clicking the roll dice button fast repeatedly then the value of total dice rolls increases immediately. And because of this some roll records could get incorrect value of roll number as there is a setTimeout of 1 second for adding the roll record in the records list. Put that increment of total dice rolls inside the setTimeout's callback function.